folder structures, routes work

+++ model/usersDb.js
@@ -14,7 +14,7 @@ const getUserDB = async (id) => {
         `, [id])
     return data
 }
-userID, firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile
+// userID, firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile
 
 const addUserDB = async (firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile) => {
     await pool.query(`
@@ -23,26 +23,17 @@ const addUserDB = async (firstName, lastName, userAge, Gender, userRole, emailAd
         `, [firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile])
 }
 
-const removeUserDB = async (id) => {
-    await pool.query(`
-        DELETE FROM users
-        WHERE userID = ?
-        `, [id])
+// userID, firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile
+
+const updateUserDB = async (firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile,id) => {
+    await pool.query(`UPDATE users SET firstName = ?, lastName = ?, userAge = ?, Gender = ?, userRole = ?,emailAdd = ?,userPass = ?,userProfile = ? WHERE userID = ?`, [firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile,id]
+
+    )
 }
 
-const updateUserDB = async (firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile,userID) => {
-    await pool.query(`
-        UPDATE users
-        SET firstName = ?, 
-        lastName = ?, 
-        userAge = ?, 
-        Gender = ?, 
-        userRole = ?,
-        emailAdd = ?
-        userPass = ?,
-        userProfile = ?
-        WHERE userID = ?
-        `, [firstName, lastName, userAge, Gender, userRole, emailAdd, userPass, userProfile,userID])
+const removeUserDB = async (id) => {
+    await pool.query(`DELETE FROM users WHERE userID = ?`, [id])
 }
+
 
 export{getUsersDB,getUserDB,addUserDB,removeUserDB,updateUserDB}
